+++ githttpxfer/git.go
@@ -4,9 +4,7 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"os/exec"
 	"path"
-	"syscall"
 )
 
 func newGit(rootPath string, binPath string, uploadPack bool, receivePack bool) *git {
@@ -45,13 +43,6 @@ func (g *git) Exists(repoPath string) bool {
 		return false
 	}
 	return true
-}
-
-func (g *git) GitCommand(repoPath string, args ...string) *exec.Cmd {
-	command := exec.Command(g.binPath, args...)
-	command.Dir = g.GetAbsolutePath(repoPath)
-	command.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	return command
 }
 
 func (g *git) GetRequestFileInfo(repoPath, filePath string) (*RequestFileInfo, error) {
