+++ githttpxfer/support_test.go
@@ -3,6 +3,8 @@ package githttpxfer
 import (
 	"net/http"
 	"net/http/httptest"
+	"os/exec"
+	"runtime"
 	"testing"
 )
 
@@ -39,6 +41,17 @@ func Test_GetServiceType(t *testing.T) {
 
 		if serviceType := getServiceType(r); tc.expected != serviceType {
 			t.Errorf("service type is not %s . result: %s", tc.expected, serviceType)
+		}
+	}
+}
+
+func Test_cleanUpProcessGroup(t *testing.T) {
+	if runtime.GOOS != "windows" {
+		cmd := exec.Command("tail", "-f", "/dev/null")
+		cmd.Start()
+		cleanUpProcessGroup(cmd)
+		if err := cmd.Wait(); err != nil && err.Error() != "signal: killed" {
+			t.Errorf("tail should have been killed, instead %v", err)
 		}
 	}
 }
